+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## __WORK IN PROGRESS__
 ### Bugfixes
 * The value IDs controlling `Start/Stop Level Change` commands in the `Multilevel Switch CC` are now also created for V1 and V2 nodes during the interview
+* The `Alarm Sensor CC` value IDs for supported sensor types are now created as soon as they are known to be supported.
 
 ## 5.2.2 (2020-10-05)
 ### Config changes