+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 -->
 
 ## __WORK IN PROGRESS__
+### Features
+* When a message should be sent to a node that is assumed to be dead, the node is now pinged first to check if it is really dead.
+
 ### Bugfixes
 * The value IDs controlling `Start/Stop Level Change` commands in the `Multilevel Switch CC` are now also created for V1 and V2 nodes during the interview
 * The `Alarm Sensor CC` value IDs for supported sensor types are now created as soon as they are known to be supported.